Always set cURL requests to CURLOPT_RETURNTRANSFER true
POST requests to the API didn't include curl_setopt($ch, CURLOPT_RETURNTRANSFER, true); Therefore when a POST request is made using HttpClient to the API, the contents of the request is directly output to the screen. This is particularly problematic with the OrderSaveObserver that watches for new orders, and sends orders to the API. In the response to a payment provider, the additional output from the POST request '[]' is appended to the API response, which screws with the Javascript methods (the response from the API can't be directly JsonDecoded.